cosmetics, move files to proper packages
