+++ app/src/main/java/edu/kmaooad/capstone23/removeAll/events/AllRemoved.java
@@ -1,0 +1,4 @@
+package edu.kmaooad.capstone23.removeAll.events;
+
+public record AllRemoved() {
+}
