--- app/src/main/java/edu/kmaooad/capstone23/removeAll/handlers/AllRemoved.java
@@ -1,4 +1,0 @@
-package edu.kmaooad.capstone23.removeAll.handlers;
-
-public record AllRemoved() {
-}
