+++ app/src/test/java/edu/kmaooad/capstone23/ban/handlers/UnbanEntityHandlerTests.java
@@ -14,7 +14,7 @@ import edu.kmaooad.capstone23.members.events.BasicMemberCreated;
 import edu.kmaooad.capstone23.orgs.commands.CreateOrg;
 import edu.kmaooad.capstone23.orgs.events.OrgCreated;
 import edu.kmaooad.capstone23.removeAll.commands.RemoveAll;
-import edu.kmaooad.capstone23.removeAll.handlers.AllRemoved;
+import edu.kmaooad.capstone23.removeAll.events.AllRemoved;
 import io.quarkus.test.junit.QuarkusTest;
 import jakarta.inject.Inject;
 import org.bson.types.ObjectId;
